+++ src/Directives/Sandbox.php
@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 
 namespace Zae\ContentSecurityPolicy\Directives;
 
@@ -14,21 +15,16 @@ namespace Zae\ContentSecurityPolicy\Directives;
  */
 class Sandbox extends Directive
 {
-    const KEY = 'sandbox';
+    public const KEY = 'sandbox';
 
-    const ALLOW_FORMS = 'allow-forms';
-    const ALLOW_MODALS = 'allow-modals';
-    const ALLOW_ORIENTATION_LOCK = 'allow-orientation-lock';
-    const ALLOW_POINTER_LOCK = 'allow-pointer-lock';
-    const ALLOW_POPUPS = 'allow-popups';
-    const ALLOW_POPUPS_TO_ESCAPE_SANDBOX = 'allow-popups-to-escape-sandbox';
-    const ALLOW_PRESENTATION = 'allow-presentation';
-    const ALLOW_SAME_ORIGIN = 'allow-same-origin';
-    const ALLOW_SCRIPTS = 'allow-scripts';
-    const ALLOW_TOP_NAVIGATION = 'allow-top-navigation';
-
-    function getKey()
-    {
-        return static::KEY;
-    }
+    public const ALLOW_FORMS = 'allow-forms';
+    public const ALLOW_MODALS = 'allow-modals';
+    public const ALLOW_ORIENTATION_LOCK = 'allow-orientation-lock';
+    public const ALLOW_POINTER_LOCK = 'allow-pointer-lock';
+    public const ALLOW_POPUPS = 'allow-popups';
+    public const ALLOW_POPUPS_TO_ESCAPE_SANDBOX = 'allow-popups-to-escape-sandbox';
+    public const ALLOW_PRESENTATION = 'allow-presentation';
+    public const ALLOW_SAME_ORIGIN = 'allow-same-origin';
+    public const ALLOW_SCRIPTS = 'allow-scripts';
+    public const ALLOW_TOP_NAVIGATION = 'allow-top-navigation';
 }