+++ src/Directives/UpgradeInsecureRequests.php
@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 
 namespace Zae\ContentSecurityPolicy\Directives;
 
@@ -12,13 +13,9 @@ namespace Zae\ContentSecurityPolicy\Directives;
  *
  * @package Zae\ContentSecurityPolicy\Directives
  */
-class UpgradeInsecureRequests extends NoValueDirective
+class UpgradeInsecureRequests extends Directive
 {
-    //todo: this directive has no value, check if that works!
-    const KEY = 'upgrade-insecure-requests';
+    use NoValueDirective;
 
-    function getKey()
-    {
-        return static::KEY;
-    }
+    public const KEY = 'upgrade-insecure-requests';
 }