+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All Notable changes to `Content Security Policy` will be documented in this file
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## Unreleased
+
+## 2.0.0 - 2019-08-28
+- PHP 7.2+
+- View Composer for $cspnonce variable
+- Strict Types
+- Autodetect Service Provider
+- Misc Refactoring
+
 ## Initial Release - 2017-01-08
 
 ### Added